controller: deal with old config schema
