+++ packages/controller/src/resources/apis/cortex.ts
@@ -39,9 +39,34 @@ import {
   ControllerConfigType
 } from "@opstrace/controller-config";
 
+import { keyIDfromPEM, log } from "@opstrace/utils";
+
 // This should really be called "Cortex API"
 // also see opstrace-prelaunch/issues/608
 // Update: said renaming effort is in progress.
+
+function generateAuthenticatorPubkeySetJSON(data_api_authn_pubkey_pem: string) {
+  // The key set is required to be a mapping between keyID (string) and
+  // PEM-encoded pubkey (string).
+  // Note: upon _continutation_, this key should be added to the existing
+  // key set.
+  const keyId = keyIDfromPEM(data_api_authn_pubkey_pem);
+  const keyset = {
+    [keyId]: data_api_authn_pubkey_pem
+  };
+
+  log.info("built authenticator key ID: %s", keyId);
+
+  // The corresponding configuration parameter value is expected to be a
+  // string, namely the above `keyset` mapping in JSON-encoded form *without
+  // literal newline chars*.
+  const tenant_api_authenticator_pubkey_set_json = JSON.stringify(keyset);
+  log.info(
+    "generated AuthenticatorPubkeySetJSON: %s",
+    tenant_api_authenticator_pubkey_set_json
+  );
+  return tenant_api_authenticator_pubkey_set_json;
+}
 
 export function CortexAPIResources(
   state: State,
@@ -76,16 +101,38 @@ export function CortexAPIResources(
   ];
 
   let cortexApiProxyEnv: V1EnvVar[];
-  if (!controllerConfig.disable_data_api_authentication) {
+  cortexApiProxyEnv = [];
+
+  if (controllerConfig.disable_data_api_authentication) {
+    cortexApiProxyCliArgs.push("-disable-api-authn");
+  }
+
+  // Support legacy controller config schema, with only
+  // `data_api_authn_pubkey_pem` being set.
+  if (controllerConfig.data_api_authn_pubkey_pem !== undefined) {
+    const pubkeypem: string = controllerConfig.data_api_authn_pubkey_pem as string;
+    if (pubkeypem.length > 0) {
+      log.info(
+        "found legacy controller config with data_api_authn_pubkey_pem parameter"
+      );
+      const keySetJSON = generateAuthenticatorPubkeySetJSON(pubkeypem);
+      cortexApiProxyEnv = [
+        {
+          name: "API_AUTHTOKEN_VERIFICATION_PUBKEY_SET",
+          value: keySetJSON
+        }
+      ];
+    }
+  } else {
+    // code path for new controller config schema, with
+    // `tenant_api_authenticator_pubkey_set_json` being set and
+    // data_api_authn_pubkey_pem not being set.
     cortexApiProxyEnv = [
       {
         name: "API_AUTHTOKEN_VERIFICATION_PUBKEY_SET",
         value: controllerConfig.tenant_api_authenticator_pubkey_set_json
       }
     ];
-  } else {
-    cortexApiProxyEnv = [];
-    cortexApiProxyCliArgs.push("-disable-api-authn");
   }
 
   collection.add(
